Module Intexable.Map_with_projection_from_key

Like a map, but the key is actually projected from the original key, and we need to provide an order on that key

module Make : functor (S : sig ... end) -> sig ... end