Module Tjr_lib_core__Intexable

Values indexed by int are quite common.

type 'a intexable = int * 'a

module Map_with_projection_from_key : sig ... end

Like a map, but the key is actually projected from the original key, and we need to provide an order on that key

module Map_make : functor (S : sig ... end) -> sig ... end