Module Tjr_lib_core__Lru_two_gen_v4.Entry

Entries in the internal maps can record that an entry has been deleted, or that an entry is present in the lower map m3;

type 'v entry = [ `Inserted of 'v | `Deleted | `Lower_some of 'v | `Lower_none ] 

type 'v entry = [

| `Inserted of 'v

| `Deleted

| `Lower_some of 'v

| `Lower_none

]

type 'v dirty_entry = [

| `Inserted of 'v

| `Deleted

]