Module Tjr_lib_core__Lru_two_gen_v4.With_explicit_cache

For some applications, we want very precise control of the cache, ie not to have the cache state hidden inside functions, but instead be referenced explicitly

module type S = sig ... end

module Make : functor (S : S) -> sig ... end