Module Tjr_lib_core__Runlevel

Simple record of the "runlevel" we have reached.

type runlvl = {

runlvl : int;

}

val l1_module_init : runlvl

val l6_main_init : runlvl

this is also the point at which main starts running

val l7_post_main_init : runlvl

the point after main init, but before main proper

val runlvl : runlvl Stdlib.ref

val set : runlvl -> unit

val is_pre_main : unit -> bool

val is_running_main : unit -> bool

val note_main_starts : unit -> unit

val note_post_main_init : unit -> unit

val can_read_config_file : unit -> bool