Module Tjr_lib_core__Tjr_show

Convert expressions to strings for debugging and logging, via sexps; uses ppx_sexp_message

You need to add ppx_sexp_message to the pps section of the dune file

See https://github.com/janestreet/ppx_sexp_message for doc

module Test : functor () sig ... end